+++ Thunderstore/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.1
+Wallride frameboost runoff now disabled by default, added config value
+
+Fixed a wallride issue
+
+New sloped rail mechanic, jumping on a rail sloped up will give a higher jump and jumping on a downward slope will give a boost of speed.
+
 
 ## 1.1.0
 New collision change, this stops almost all instances of clipping through walls
@@ -16,14 +23,18 @@ Super sliding on carpets has been improved
 
 Hard landings have been changed, if in a movestyle you will no longer have a hard landing, if on your feet you will only have a hard landing if moving past your default maximum speed and having been in the air for 1.5 seconds.
 
+
 ## 1.0.3
 whoops, fixed the vert stuff
+
 
 ## 1.0.2
 More Slopcrew fixes
 
+
 ## 1.0.1
 Fixed the perfect manual not working with Slopcrew
+
 
 ## 1.0.0
 Initial release